requirements.txt
dash>=2.17
plotly>=5.22
pandas>=2.2
playwright>=1.44
httpx>=0.27
python-dotenv>=1.0
